+++ GreenTitan/Functions.h
@@ -4,7 +4,7 @@
 #include "Arduino.h"
 #include "Defines.h"
 
-float angleBetweenPoints(long x1, long y1, long x2, long y2);
+float angleBetweenPoints(int x1, int y1, int x2, int y2);
 float NormalizeAngle(float angle);
 float absAngle(float angle);
 float ShortestRotation(float targetAngle, float currentAngle);
